+++ src/ui/about.rs
@@ -107,7 +107,20 @@ impl SimpleComponent for AboutDialog {
                 "<p>Fixed</p>",
 
                 "<ul>",
-                    "<li>Fixed main patch applying</li>",
+                    "<li>Fixed telemetry disabling</li>",
+                "</ul>",
+
+                "<p>Changed</p>",
+
+                "<ul>",
+                    "<li>Updated Italian</li>",
+                    "<li>Updated Hungarian</li>",
+                    "<li>Updated Japanese</li>",
+                    "<li>Updated Indonesian</li>",
+                    "<li>Updated Spanish</li>",
+                    "<li>Updated Turkish</li>",
+                    "<li>Updated Chinese</li>",
+                    "<li>Disabled patch applying for new installations</li>",
                 "</ul>",
             ].join("\n"),